chore(developer): move compfile.h to kmcmpdll
Moves compfile.h to kmcmpdll path and removes unnecessary dependencies on it.
