+++ developer/src/kmcmpdll/kmcmpdll.vcxproj
@@ -290,7 +290,7 @@
     <ClInclude Include="$(KEYMAN_ROOT)\developer\src\ext\json\json.hpp" />
     <ClInclude Include="$(KEYMAN_ROOT)\windows\src\global\inc\certificate_check.h" />
     <ClInclude Include="comperr.h" />
-    <ClInclude Include="$(KEYMAN_ROOT)\windows\src\global\inc\Compfile.h" />
+    <ClInclude Include="compfile.h" />
     <ClInclude Include="$(KEYMAN_ROOT)\windows\src\global\inc\Compiler.h" />
     <ClInclude Include="CasedKeys.h" />
     <ClInclude Include="CharToKeyConversion.h" />
