+++ developer/src/test/auto/kmcomp-x64-structures/cppstructsize/cppstructsize.vcxproj
@@ -159,7 +159,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\..\..\..\..\windows\src\global\inc\Compfile.h" />
+    <ClInclude Include="../../../../kmcmpdll/compfile.h" />
     <ClInclude Include="pch.h" />
   </ItemGroup>
   <ItemGroup>
