+++ developer/src/test/auto/kmcomp-x64-structures/cppstructsize/cppstructsize.vcxproj.filters
@@ -18,7 +18,7 @@
     <ClInclude Include="pch.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\..\..\global\inc\Compfile.h">
+    <ClInclude Include="..\..\..\..\kmcmpdll\compfile.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
